STYLE: Convert 'enum' to 'enum class' objects with print enum function
Improve use of enumerations to new features in C++11.  A set of
refactorings to move from "c enumerations" to "strongly typed
enumerations" is included.

See: http://www.stroustrup.com/C++11FAQ.html#enum

The ITKSmoothing module now has a library that must be linked, so the
module dependancies must be moved from COMPILE_DEPENDS to DEPENDS.

Proper ENABLE_SHARED, and ITK{MODULENAME}_EXPORT with
`#include "ITK{MODULENAME}_Export.h"` instrumentation for
the ostream operators used with the enumerations are added.

Co-Authored by: Mathew J. Seng <mathewseng@gmail.com>

+++ Modules/Core/SpatialObjects/include/itkDTITubeSpatialObjectPoint.hxx
@@ -53,7 +53,7 @@ DTITubeSpatialObjectPoint< TPointDimension >
 ::TranslateEnumToChar(FieldEnumType name) const
 {
   // Do the translation
-  switch ( name )
+  switch (static_cast<int>(name) )
     {
     case 0:
       return std::string("FA");
